requirements.txt
gherkin-official==29.0.0
greenlet==3.1.1
iniconfig==2.0.0
Mako==1.3.9
MarkupSafe==3.0.2
packaging==24.2
parse==1.20.2
parse_type==0.6.4
playwright==1.50.0
pluggy==1.5.0
pyee==12.1.1
pytest==8.3.5
pytest-bdd==8.1.0
pytest-env==1.1.5
python-dotenv==1.0.1
six==1.17.0
typing_extensions==4.12.2
